[Refactor] refactor split agg rule
