+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/logical/LogicalAggregationOperator.java
@@ -59,17 +59,6 @@ public class LogicalAggregationOperator extends LogicalOperator {
     // but for single distinct function, partitionByColumns are not same with groupingKeys
     private List<ColumnRefOperator> partitionByColumns;
 
-    // @todo: refactor it, SingleDistinctFunctionPos depend on the map's order, but the order is not fixed
-    // When generate plan fragment, we need this info.
-    // For SQL: select count(distinct id_bigint), sum(id_int) from test_basic;
-    // In the distinct local (update serialize) agg stage:
-    //|   5:AGGREGATE (update serialize)                                                      |
-    //|   |  output: count(<slot 13>), sum(<slot 16>)                                         |
-    //|   |  group by:                                                                        |
-    // count function is update function, but sum is merge function
-    // if singleDistinctFunctionPos is -1, means no single distinct function
-    private int singleDistinctFunctionPos = -1;
-
     private DataSkewInfo distinctColumnDataSkew = null;
 
     // If the AggType is not GLOBAL, it means we have split the agg hence the isSplit should be true.
@@ -78,7 +67,7 @@ public class LogicalAggregationOperator extends LogicalOperator {
     public LogicalAggregationOperator(AggType type,
                                       List<ColumnRefOperator> groupingKeys,
                                       Map<ColumnRefOperator, CallOperator> aggregations) {
-        this(type, groupingKeys, groupingKeys, aggregations, false, -1, -1, null);
+        this(type, groupingKeys, groupingKeys, aggregations, false, -1, null);
     }
 
     public LogicalAggregationOperator(
@@ -87,7 +76,6 @@ public class LogicalAggregationOperator extends LogicalOperator {
             List<ColumnRefOperator> partitionByColumns,
             Map<ColumnRefOperator, CallOperator> aggregations,
             boolean isSplit,
-            int singleDistinctFunctionPos,
             long limit,
             ScalarOperator predicate) {
         super(OperatorType.LOGICAL_AGGR, limit, predicate, null);
@@ -96,13 +84,11 @@ public class LogicalAggregationOperator extends LogicalOperator {
         this.partitionByColumns = partitionByColumns;
         this.aggregations = ImmutableMap.copyOf(aggregations);
         this.isSplit = !type.isGlobal() || isSplit;
-        this.singleDistinctFunctionPos = singleDistinctFunctionPos;
     }
 
     private LogicalAggregationOperator() {
         super(OperatorType.LOGICAL_AGGR);
         this.isSplit = false;
-        this.singleDistinctFunctionPos = -1;
     }
 
     public AggType getType() {
@@ -123,10 +109,6 @@ public class LogicalAggregationOperator extends LogicalOperator {
 
     public void setOnlyLocalAggregate() {
         isSplit = false;
-    }
-
-    public int getSingleDistinctFunctionPos() {
-        return singleDistinctFunctionPos;
     }
 
     public List<ColumnRefOperator> getPartitionByColumns() {
@@ -236,7 +218,7 @@ public class LogicalAggregationOperator extends LogicalOperator {
             return false;
         }
         LogicalAggregationOperator that = (LogicalAggregationOperator) o;
-        return isSplit == that.isSplit && singleDistinctFunctionPos == that.singleDistinctFunctionPos &&
+        return isSplit == that.isSplit &&
                 type == that.type && Objects.equals(aggregations, that.aggregations) &&
                 Objects.equals(groupingKeys, that.groupingKeys) &&
                 Objects.equals(partitionByColumns, that.partitionByColumns);
@@ -244,8 +226,7 @@ public class LogicalAggregationOperator extends LogicalOperator {
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), type, isSplit, aggregations, groupingKeys, partitionByColumns,
-                singleDistinctFunctionPos);
+        return Objects.hash(super.hashCode(), type, isSplit, aggregations, groupingKeys, partitionByColumns);
     }
 
     public static Builder builder() {
@@ -277,7 +258,6 @@ public class LogicalAggregationOperator extends LogicalOperator {
             builder.partitionByColumns = aggregationOperator.partitionByColumns;
             builder.aggregations = aggregationOperator.aggregations;
             builder.isSplit = aggregationOperator.isSplit;
-            builder.singleDistinctFunctionPos = aggregationOperator.singleDistinctFunctionPos;
             builder.distinctColumnDataSkew = aggregationOperator.distinctColumnDataSkew;
             return this;
         }
@@ -306,11 +286,6 @@ public class LogicalAggregationOperator extends LogicalOperator {
         public Builder setPartitionByColumns(
                 List<ColumnRefOperator> partitionByColumns) {
             builder.partitionByColumns = partitionByColumns;
-            return this;
-        }
-
-        public Builder setSingleDistinctFunctionPos(int singleDistinctFunctionPos) {
-            builder.singleDistinctFunctionPos = singleDistinctFunctionPos;
             return this;
         }
