+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalHashAggregateOperator.java
@@ -53,17 +53,6 @@ public class PhysicalHashAggregateOperator extends PhysicalOperator {
     private final List<ColumnRefOperator> partitionByColumns;
     private final Map<ColumnRefOperator, CallOperator> aggregations;
 
-    // @todo: refactor it, SingleDistinctFunctionPos depend on the map's order, but the order is not fixed
-    // When generate plan fragment, we need this info.
-    // For select count(distinct id_bigint), sum(id_int) from test_basic;
-    // In the distinct local (update serialize) agg stage:
-    // |   5:AGGREGATE (update serialize)                                                      |
-    //|   |  output: count(<slot 13>), sum(<slot 16>)                                         |
-    //|   |  group by:                                                                        |
-    // count function is update function, but sum is merge function
-    // if singleDistinctFunctionPos is -1, means no single distinct function
-    private final int singleDistinctFunctionPos;
-
     // The flag for this aggregate operator has split to
     // two stage aggregate or three stage aggregate
     private final boolean isSplit;
@@ -81,7 +70,6 @@ public class PhysicalHashAggregateOperator extends PhysicalOperator {
                                          List<ColumnRefOperator> groupBys,
                                          List<ColumnRefOperator> partitionByColumns,
                                          Map<ColumnRefOperator, CallOperator> aggregations,
-                                         int singleDistinctFunctionPos,
                                          boolean isSplit,
                                          long limit,
                                          ScalarOperator predicate,
@@ -91,7 +79,6 @@ public class PhysicalHashAggregateOperator extends PhysicalOperator {
         this.groupBys = groupBys;
         this.partitionByColumns = partitionByColumns;
         this.aggregations = aggregations;
-        this.singleDistinctFunctionPos = singleDistinctFunctionPos;
         this.isSplit = isSplit;
         this.limit = limit;
         this.predicate = predicate;
@@ -132,12 +119,8 @@ public class PhysicalHashAggregateOperator extends PhysicalOperator {
         return partitionByColumns;
     }
 
-    public boolean hasSingleDistinct() {
-        return singleDistinctFunctionPos > -1;
-    }
-
-    public int getSingleDistinctFunctionPos() {
-        return singleDistinctFunctionPos;
+    public boolean hasRemoveDistinctFunc() {
+        return aggregations.values().stream().anyMatch(CallOperator::isRemovedDistinct);
     }
 
     public boolean isSplit() {
@@ -157,7 +140,7 @@ public class PhysicalHashAggregateOperator extends PhysicalOperator {
 
     public String getNeededPreaggregationMode() {
         String mode = ConnectContext.get().getSessionVariable().getStreamingPreaggregationMode();
-        if (canUseStreamingPreAgg() && (type.isDistinctLocal() || hasSingleDistinct())) {
+        if (canUseStreamingPreAgg() && (type.isDistinctLocal() || hasRemoveDistinctFunc())) {
             mode = SessionVariableConstants.FORCE_PREAGGREGATION;
         }
         return mode;
