+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleType.java
@@ -31,6 +31,9 @@ public enum RuleType {
     TF_CBO_TABLE_PRUNE_RULE,
     TF_LIMIT_TABLETS_PRUNE,
     TF_SPLIT_AGGREGATE,
+    TF_SPLIT_WITHOUT_DISTINCT_AGGREGATE,
+    TF_SPLIT_DISTINCT_AGGREGATE,
+
     TF_SPLIT_TOPN,
     TF_SPLIT_SCAN_OR,
     TF_PUSH_DOWN_JOIN_AGG,
