+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/mv/MaterializedViewRewriter.java
@@ -203,7 +203,6 @@ public class MaterializedViewRewriter extends OptExpressionVisitor<OptExpression
                 aggregationOperator.getPartitionByColumns(),
                 newAggMap,
                 aggregationOperator.isSplit(),
-                aggregationOperator.getSingleDistinctFunctionPos(),
                 aggregationOperator.getLimit(),
                 aggregationOperator.getPredicate()), optExpression.getInputs());
     }
