+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/tree/lowcardinality/DecodeContext.java
@@ -329,13 +329,15 @@ class DecodeContext {
                 Function fn = Expr.getBuiltinFunction(call.getFnName(), argTypes, Function.CompareMode.IS_SUPERTYPE_OF);
                 // min/max function: will rewrite all stage, return type is dict type
                 if (FunctionSet.MAX.equals(call.getFnName()) || FunctionSet.MIN.equals(call.getFnName())) {
-                    return new CallOperator(call.getFnName(), fn.getReturnType(), newChildren, fn, call.isDistinct());
+                    return new CallOperator(call.getFnName(), fn.getReturnType(), newChildren, fn,
+                            call.isDistinct(), call.isRemovedDistinct());
                 } else if (FunctionSet.COUNT.equals(call.getFnName()) ||
                         FunctionSet.MULTI_DISTINCT_COUNT.equals(call.getFnName()) ||
                         FunctionSet.APPROX_COUNT_DISTINCT.equals(call.getFnName())) {
                     // count, count_distinct, approx_count_distinct:
                     // return type don't update
-                    return new CallOperator(call.getFnName(), call.getType(), newChildren, fn, call.isDistinct());
+                    return new CallOperator(call.getFnName(), call.getType(), newChildren, fn,
+                            call.isDistinct(), call.isRemovedDistinct());
                 }
             }
 
@@ -348,7 +350,8 @@ class DecodeContext {
                 argTypes[i] = newChildren.get(i).getType();
             }
             Function fn = Expr.getBuiltinFunction(call.getFnName(), argTypes, Function.CompareMode.IS_SUPERTYPE_OF);
-            return new CallOperator(call.getFnName(), fn.getReturnType(), newChildren, fn, call.isDistinct());
+            return new CallOperator(call.getFnName(), fn.getReturnType(), newChildren, fn,
+                    call.isDistinct(), call.isRemovedDistinct());
         }
     }
