+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/validate/TypeChecker.java
@@ -86,8 +86,7 @@ public class TypeChecker implements PlanValidator.Checker {
         @Override
         public Void visitLogicalAggregate(OptExpression optExpression, Void context) {
             LogicalAggregationOperator operator = (LogicalAggregationOperator) optExpression.getOp();
-            checkAggCall(operator.getAggregations(), operator.getType(), operator.isSplit(),
-                    operator.getSingleDistinctFunctionPos());
+            checkAggCall(operator.getAggregations(), operator.getType(), operator.isSplit());
             visit(optExpression, context);
             return null;
         }
@@ -104,8 +103,7 @@ public class TypeChecker implements PlanValidator.Checker {
         @Override
         public Void visitPhysicalHashAggregate(OptExpression optExpression, Void context) {
             PhysicalHashAggregateOperator operator = (PhysicalHashAggregateOperator) optExpression.getOp();
-            checkAggCall(operator.getAggregations(), operator.getType(), operator.isSplit(),
-                    operator.getSingleDistinctFunctionPos());
+            checkAggCall(operator.getAggregations(), operator.getType(), operator.isSplit());
             visit(optExpression, context);
             return null;
         }
@@ -147,10 +145,8 @@ public class TypeChecker implements PlanValidator.Checker {
         }
 
 
-        private void checkAggCall(Map<ColumnRefOperator, CallOperator> aggregations, AggType aggType, boolean isSplit,
-                                  int singleDistinctFunctionPos) {
+        private void checkAggCall(Map<ColumnRefOperator, CallOperator> aggregations, AggType aggType, boolean isSplit) {
             int index = -1;
-            boolean hasSingleDistinct = singleDistinctFunctionPos != - 1;
             for (Map.Entry<ColumnRefOperator, CallOperator> entry : aggregations.entrySet()) {
                 index++;
                 ColumnRefOperator outputCol = entry.getKey();
@@ -158,16 +154,19 @@ public class TypeChecker implements PlanValidator.Checker {
                 boolean isMergeAggFn = false;
                 switch (aggType) {
                     case LOCAL:
-                        isMergeAggFn = isSplit && hasSingleDistinct && (index != singleDistinctFunctionPos);
-                        break;
                     case GLOBAL:
-                        isMergeAggFn = isSplit && (!hasSingleDistinct || index != singleDistinctFunctionPos);
+                        // In 3-phase agg, the global agg should set these not distinct agg callOperator to mergeAggFn.
+                        // Sometimes we may further split the top global agg from 3-phase agg into a two phase agg
+                        // like local -> distinct global -> global to local -> distinct global -> local -> global.
+                        // we also need to set mergeAggFn flag into the second local agg for those not distinct
+                        // agg callOperator.
+                        isMergeAggFn = isSplit && !aggCall.isRemovedDistinct();
                         break;
                     case DISTINCT_GLOBAL:
                         isMergeAggFn = true;
                         break;
                     case DISTINCT_LOCAL:
-                        isMergeAggFn = index != singleDistinctFunctionPos;
+                        isMergeAggFn = !aggCall.isRemovedDistinct();
                 }
 
                 if (aggType.isGlobal()) {
